Done Inventory

+++ app/Http/Controllers/StockinController.php
@@ -38,6 +38,24 @@ class StockinController extends Controller
             ]);
         }
 
+        $date_today = Carbon::now()->toDateString();
+
+
+        // foreach ($categories as $index => $category) {
+
+        //     $is_exist = Inventory::where('code', $codes[$index])
+        //         ->whereDate('created_at', $date_today)
+        //         ->exists();
+
+        //     if ($is_exist) {
+        //         return response()->json([
+        //             'status' => 404,
+        //             'message' => 'You are not allowed to restock this code since you just inserted it today to the masterlist. In order to proceed please remove this code and update it later or by the next day.'
+        //         ]);
+        //     }
+        // }
+
+
         $data = [];
         foreach ($categories as $index => $category) {
             $data[] = [
@@ -48,9 +66,19 @@ class StockinController extends Controller
                 'price' => $prices[$index],
                 'stocks' => $stocks[$index],
                 'total_amount' => $prices[$index] * $stocks[$index],
-                'created_at' => now(),
+                'created_at' => '2024-05-27',
                 'updated_at' => now(),
             ];
+
+            $inventory = Inventory::where('code', $codes[$index])->first();
+
+            $end_env = ($inventory->stockin + $stocks[$index]) - $inventory->stockout;
+            $total_amount = $end_env * $inventory->price;
+
+            Inventory::where('code', $codes[$index])->update([
+                'end_inv' =>  $end_env,
+                'total_amount' => $total_amount
+            ]);
         }
 
         try {
@@ -81,14 +109,32 @@ class StockinController extends Controller
 
     function destroy($id)
     {
+
         $stocks = Stockin::findOrFail($id);
-        $stocks->delete();
 
         if (!$stocks) {
             return redirect()->back()->with('error', 'Failed to save record');
         }
 
-        return redirect()->route('Admins.InventoryStockList')->with('success', 'Successfully deleted item');
+        $is_beginning = Inventory::where('code', $stocks->code)
+            ->whereDate('created_at', $stocks->created_at)
+            ->exists();
+
+        if ($is_beginning) {
+            return redirect()->route('Admins.InventoryStockList')->with('warning', 'You are not allowed to delete this item since your beginning inventory is anchored to this code but, if you meant to delete or move it to waste then go to checklist table and do the operation needed to be done.');
+        } else {
+            $res = Inventory::where('code', $stocks->code)->first();
+
+            $newEnv = $res->end_inv - $stocks->stocks;
+
+
+            Inventory::where('code', $stocks->code)->update([
+                'end_inv' => $newEnv,
+                'total_amount' => $newEnv * $res->price,
+            ]);
+            $stocks->delete();
+            return redirect()->route('Admins.InventoryStockList')->with('success', 'Successfully delete item');
+        }
     }
 
 
@@ -113,18 +159,52 @@ class StockinController extends Controller
         $price = (float) $request->price;
         $total = (float) ($request->stocks * $price);
 
-        $update = Stockin::where('id', $request->id)->update(
-            [
-                'total_amount' => $total,
-                'stocks' => $request->stocks,
-                'created_at' => now(),
-            ]
-        );
+
+        $currentYear = Carbon::now()->year;
+        $currentMonth = Carbon::now()->month;
+        $totalStockins = 0;
+
+        $stocks = Stockin::where('id', $request->id)->first();
+        $inventory = Inventory::where('code', $stocks->code)->first();
+
+        if ($request->stocks != $request->old_stocks) {
+            $stockins = Stockin::where('code', $stocks->code)
+                ->whereYear('created_at', $currentYear)
+                ->whereMonth('created_at', $currentMonth)
+                ->get();
 
 
-        if (!$update) {
-            return redirect()->back()->with('error', 'Failed to update item');
+
+
+            foreach ($stockins as $stocks) {
+                $totalStockins  += $stocks->stocks;
+            }
+
+            // dd((($request->stocks - $request->old_stocks) + $totalStockins) - $inventory->stockout);
+
+            $newStocks = (($request->stocks - $request->old_stocks) + $totalStockins) - $inventory->stockout;
+
+            // dd( 'New Stoacaks: '. $request->stocks . ' old stocks: ' . $request->old_stocks . ' New stocks: '.  $newStocks . ' Stockut: '. $inventory->stockout . ' Ending :' .(($request->stocks - $request->old_stocks) + $totalStockins) - $inventory->stockout); // 3
+
+            $update = Inventory::where('code', $stocks->code)->update(
+                [
+                    'end_inv' => (($request->stocks - $request->old_stocks) + $totalStockins) - $inventory->stockout,
+                    'total_amount' => $total,
+                    'stockin' => (($request->stocks - $request->old_stocks) + $totalStockins) - $inventory->stockout,
+                ]
+            );
+
+            $update = Stockin::where('id', $request->id)->update(
+                [
+                    'total_amount' => $total,
+                    'stocks' => $request->stocks,
+                    'updated_at' => now(),
+                ]
+            );
+
+            return redirect()->route('Admins.InventoryStockList')->with('success', 'Successfully update item');
+        } else {
+            return redirect()->route('Admins.InventoryStockList')->with('warning', 'NO changes has been made');
         }
-        return redirect()->route('Admins.InventoryStockList')->with('success', 'Successfully update item');
     }
 }
